Move Functionality to useContactInfo
