+++ client/components/index.tsx
@@ -10,9 +10,11 @@ export * from './CheckBoxExplain';
 // HOCs
 export * from './withAuth';
 
-// From
+// Form
 export * from './form';
 
-// COnstants
-
+// Constants
 export * from '../constants';
+
+// Contexts
+export * from '../contexts';
